Refactor pattern lib into blueprint
