+++ app/patternlib/templates/pattern_lib/1-styleguide/3-grid-system.html
