+++ app/patternlib/templates/pattern_lib/2-patterns/1-questions.html
